backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
razorpay==2.0.1
psycopg2-binary==2.9.10
boto3==1.42.83
django-storages==1.14.6
gunicorn==23.0.0
Pillow==12.2.0
